requirements.txt
streamlit
opencv-python
numpy
matplotlib
reportlab
Pillow